Add files via upload
This is my first hands-on experience with the machine learning project. In this project, we are predicting the prices of the house in the California (USA). The guidance throughout the project is taken from the book(Aurélien-Géron-Hands-On-Machine-Learning-with-Scikit-Learn-Keras-and-Tensorflow) which I found very helpful. In this project, I have used Sklearn library of the python programming language which is so great and helping us in doing difficult tasks very easily without doing too much effort. This project gave me a good idea what a machine learning project exactly looks like. It have gone through various phases such as data collecting, analyzing the data, data cleaning, data preparation, extracting various important features, splitting the data in such a way that it is representative etc. Then after getting cleaned and prepared data, training different Ml models provided by Sklearn and checking the performance of each model that how accurately the model is predicting the house prices and then best model is chosen. Then, the important task "hyperparameter tuning" for that selected model is performed and best hyperparameters are used and them again training the model with those parameters. After finalizing the model, the evaluation is performed on the final model using the test data.
Although, the model built is not so much accurate but it was very great experience and learning.